+++ CHANGELOG.md
@@ -9,6 +9,11 @@
         - `ProgressStartDefault`
 - Allow to create inactive progress bar
     - Add `ProgressBar.inactive` function
+- Add custom tag
+    - `number` to style numbers
+    - `u` for underline
+    - `b` for bold
+    - `i` for italic
 
 ## 4.1.0 - 2022-10-19
 - Show errors with more then 2 lines just in red foreground, without a block